// src/config/filters.generated.ts
export const FILTER_DEFS: FilterDef[] = [
  {
    "key": "sites.site_name",
    "domain": "sites",
    "field": "site_name",
    "label": "Site Name",
    "match": "exact",
    "dynamic": true
  },
// ...
];